Fixing agenda functions & status show only from today

+++ resources/js/app.js
@@ -32,6 +32,9 @@ document.addEventListener('DOMContentLoaded', function () {
             select: function (arg) {
                 var title = prompt('Afspraak naam:');
                 var description = prompt('Beschrijving:');
+                if (!description) {
+                    description = "-";
+                }
                 if (title) {
                     $.ajax({
                         url: "http://127.0.0.1:8000/events",
@@ -45,7 +48,6 @@ document.addEventListener('DOMContentLoaded', function () {
                         },
                         type: "POST",
                         success: function (res) {
-                            // calendar.addEvent(res);
                             calendar.refetchEvents();
                         },
                         error: function (res) {
@@ -89,55 +91,18 @@ document.addEventListener('DOMContentLoaded', function () {
                 });
             },
             eventClick: function (info) {
-                // Populate the modal with event data
-                $('#eventTitle').val(info.event.title);
-                $('#eventDescription').val(info.event.extendedProps.description);
-                $('#eventUser').val(info.event.extendedProps.user);
-
-                // Show the modal
-                $('#eventModal').modal('show');
-
-                // Save changes
-                $('#saveEventBtn').off('click').on('click', function () {
-                    $.ajax({
-                        url: `http://127.0.0.1:8000/events/${info.event.id}`,
-                        headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') },
-                        data: {
-                            id: info.event.id,
-                            title: $('#eventTitle').val(),
-                            description: $('#eventDescription').val(),
-                            start: moment(info.event.start).format(DATEFORMAT),
-                            end: info.event.end ? moment(info.event.end).format(DATEFORMAT) : moment(info.event.start).format(DATEFORMAT),
-                            all_day: Number(info.event.allDay),
-                        },
-                        type: "PATCH",
-                        success: function (res) {
-                            info.event.setProp('title', $('#eventTitle').val());
-                            info.event.setExtendedProp('description', $('#eventDescription').val());
-                            $('#eventModal').modal('hide');
-                            calendar.refetchEvents();
-                        },
-                        error: function (res) {
-                            console.log(res);
-                        }
-                    });
-                });
-
-                // Delete event
-                $('#deleteEventBtn').off('click').on('click', function () {
-                    $.ajax({
-                        url: `http://127.0.0.1:8000/events/${info.event.id}`,
-                        headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') },
-                        type: "DELETE",
-                        success: function (res) {
-                            info.event.remove();
-                            $('#eventModal').modal('hide');
-                        },
-                        error: function (res) {
-                            console.log(res);
-                        }
-                    });
-                });
+                if (confirm("Are you sure you want to delete this event?")) {
+                $.ajax({
+                    url: `http://127.0.0.1:8000/events/${info.event.id}`,
+                    headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') },
+                    type: "DELETE",
+                    success: function (res) {
+                        calendar.refetchEvents();
+                    },
+                    error: function (res) {
+                        console.log(res);
+                    }
+                });}
             },
             eventSources: '/events', // Endpoint to fetch existing events
             headerToolbar: {
